+++ src/transport/common/tx.c
@@ -27,7 +27,7 @@ void __unsafe_z_prepare_wbuf(_z_wbuf_t *buf, int is_streamed) {
     _z_wbuf_reset(buf);
 
     if (is_streamed == 1) {
-        for (size_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
+        for (uint8_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
             _z_wbuf_put(buf, 0, i);
         }
         _z_wbuf_set_wpos(buf, _Z_MSG_LEN_ENC_SIZE);
@@ -42,8 +42,8 @@ void __unsafe_z_prepare_wbuf(_z_wbuf_t *buf, int is_streamed) {
 void __unsafe_z_finalize_wbuf(_z_wbuf_t *buf, int is_streamed) {
     if (is_streamed == 1) {
         size_t len = _z_wbuf_len(buf) - _Z_MSG_LEN_ENC_SIZE;
-        for (size_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
-            _z_wbuf_put(buf, (uint8_t)((len >> (size_t)8 * i) & (uint8_t)0xFF), i);
+        for (uint8_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
+            _z_wbuf_put(buf, (uint8_t)((len >> (uint8_t)8 * i) & (uint8_t)0xFF), i);
         }
     }
 }
@@ -114,7 +114,7 @@ int _z_link_send_t_msg(const _z_link_t *zl, const _z_transport_message_t *t_msg)
     uint16_t mtu = (zl->_mtu < Z_BATCH_SIZE_TX) ? zl->_mtu : Z_BATCH_SIZE_TX;
     _z_wbuf_t wbf = _z_wbuf_make(mtu, 0);
     if (_Z_LINK_IS_STREAMED(zl->_capabilities)) {
-        for (size_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
+        for (uint8_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
             _z_wbuf_put(&wbf, 0, i);
         }
         _z_wbuf_set_wpos(&wbf, _Z_MSG_LEN_ENC_SIZE);
@@ -128,8 +128,8 @@ int _z_link_send_t_msg(const _z_link_t *zl, const _z_transport_message_t *t_msg)
     // Write the message legnth in the reserved space if needed
     if (_Z_LINK_IS_STREAMED(zl->_capabilities)) {
         size_t len = _z_wbuf_len(&wbf) - _Z_MSG_LEN_ENC_SIZE;
-        for (size_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
-            _z_wbuf_put(&wbf, (uint8_t)((len >> (size_t)8 * i) & (uint8_t)0xFF), i);
+        for (uint8_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
+            _z_wbuf_put(&wbf, (uint8_t)((len >> (uint8_t)8 * i) & (uint8_t)0xFF), i);
         }
     }
