+++ src/transport/unicast/link/rx.c
@@ -44,8 +44,8 @@ void _z_unicast_recv_t_msg_na(_z_transport_unicast_t *ztu, _z_transport_message_
         }
 
         size_t len = 0;
-        for (int i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
-            len |= _z_zbuf_read(&ztu->_zbuf) << (i * 8);
+        for (uint8_t i = 0; i < _Z_MSG_LEN_ENC_SIZE; i++) {
+            len |= _z_zbuf_read(&ztu->_zbuf) << (i * (uint8_t)8);
         }
 
         _Z_DEBUG(">> \t msg len = %zu\n", len);
